🔨 Refactor: 캐러셀 데이터 분리
